Modularize code by extracting genre utilities
